Revert "design:tempPart1을 파트별집중학습 컨셉에 맞게 변경"
This reverts commit 64531e4e858e7819b65cd856c9a7982d2c396307.
